#14 move logic out of goferbot.js
