+++ lib/takeOrder.js
@@ -1,10 +1,38 @@
-var logic = require('./witlogic.js');
-var dbmanager = require('./dbmanager.js');
 var EventEmitter = require('events');
 var Wit = require('node-wit').Wit;
+var logic = require('./witlogic.js');
+var dbmanager = require('./dbmanager.js');
+
 var wit = new Wit(process.env.WIT_TOKEN, logic.actions);
 
-module.exports = function takeOrder(convo) {
+module.exports = function (bot) {
+  // new CronJob('00 45 09 * * 1-5', call, null, true, 'America/Los_Angeles');
+  var teamId = bot.team_info.id;
+  dbmanager.customerList(teamId)
+  .then(function(customers) {
+    customers.forEach(function(customer) {
+      bot.startPrivateConversation(customer, ordering);
+    });
+  });
+}
+
+function ordering(error, convo) {
+  convo.sessionId = convo.sessionId || uuid.v1();
+  convo.context = {};
+  var slackApi = convo.task.bot.api;
+  var userInfo = { user: convo.source_message.user };
+  slackApi.users.info(userInfo, function(err, ping) {
+    if (!err && ping.ok) {
+      convo.user = ping.user.profile;
+      convo.user.name = ping.user.name;
+      convo.user.id = ping.user.id;
+      var greeting = 'Good morning! I\'m taking coffee orders right now, would you like anything?'
+      convo.ask(greeting, takeOrder(convo).askProcessing);
+    }
+  });
+}
+
+function takeOrder(convo) {
   convo.events = new EventEmitter();
   convo.events.on('typing', typing);
   convo.events.on('merge', merge);
@@ -70,28 +98,38 @@ module.exports = function takeOrder(convo) {
   }
 
   function action(actionName) {
-    var beverage = {
-      bev: convo.context.bev,
-      size: convo.context.size
-    };
-    var ticket = {
-      user: convo.user,
-      beverage: beverage
-    };
+    if (actionName == 'placeOrder') {
+      var beverage = {
+        bev: convo.context.bev,
+        size: convo.context.size
+      };
+      var ticket = {
+        user: convo.user,
+        beverage: beverage
+      };
 
-    function cb() {
-      dbmanager.add(ticket).then(function(results) {
-        console.log('successfully added');
-      });
+      function cb() {
+        dbmanager.add(ticket).then(function(results) {
+          console.log('successfully added');
+        });
+      }
+
+      logic.actions[actionName](convo.sessionId,convo.context, cb);
+      wit.converse(
+        convo.sessionId,
+        undefined,
+        convo.context,
+        witProcessing
+      );
     }
-
-    logic.actions[actionName](convo.sessionId,convo.context, cb);
-    wit.converse(
-      convo.sessionId,
-      undefined,
-      convo.context,
-      witProcessing
-    );
+    else {
+      wit.converse(
+        convo.sessionId,
+        undefined,
+        convo.context,
+        witProcessing
+      );
+    }
   }
 
   function end() {
